Moved android tests to data
