+++ data/src/androidTest/java/com/moonlightbutterfly/cryptohub/TestRepositoryModule.kt
@@ -1,29 +1,28 @@
-package com.moonlightbutterfly.cryptohub.di
+package com.moonlightbutterfly.cryptohub
 
 import android.content.Context
 import androidx.work.WorkManager
 import com.firebase.ui.auth.AuthUI
 import com.google.firebase.auth.FirebaseAuth
+import com.moonlightbutterfly.cryptohub.assets.FakeCryptoAssetsDataSourceImpl
+import com.moonlightbutterfly.cryptohub.collections.FakeUserCollectionsLocalDataSourceImpl
+import com.moonlightbutterfly.cryptohub.collections.FakeUserCollectionsRemoteDataSourceImpl
 import com.moonlightbutterfly.cryptohub.data.assets.CryptoAssetsDataSource
 import com.moonlightbutterfly.cryptohub.data.assets.CryptoAssetsRepository
-import com.moonlightbutterfly.cryptohub.data.assets.FakeCryptoAssetsDataSourceImpl
-import com.moonlightbutterfly.cryptohub.data.collections.FakeUserCollectionsLocalDataSourceImpl
-import com.moonlightbutterfly.cryptohub.data.collections.FakeUserCollectionsRemoteDataSourceImpl
 import com.moonlightbutterfly.cryptohub.data.collections.UserCollectionsLocalDataSource
 import com.moonlightbutterfly.cryptohub.data.collections.UserCollectionsRemoteDataSource
 import com.moonlightbutterfly.cryptohub.data.collections.UserCollectionsRepository
-import com.moonlightbutterfly.cryptohub.data.localpreferences.FakeLocalPreferencesDataSourceImpl
 import com.moonlightbutterfly.cryptohub.data.localpreferences.LocalPreferencesDataSource
 import com.moonlightbutterfly.cryptohub.data.localpreferences.LocalPreferencesRepository
 import com.moonlightbutterfly.cryptohub.data.notifications.Notifier
-import com.moonlightbutterfly.cryptohub.data.notifications.NotifierImpl
-import com.moonlightbutterfly.cryptohub.data.signin.FakeFirebaseSignInHandler
-import com.moonlightbutterfly.cryptohub.data.signin.FirebaseSignInHandler
-import com.moonlightbutterfly.cryptohub.data.user.FakeUserDataSourceImpl
-import com.moonlightbutterfly.cryptohub.data.user.FirebaseAuthDataProvider
-import com.moonlightbutterfly.cryptohub.data.user.FirebaseAuthDataProviderImpl
 import com.moonlightbutterfly.cryptohub.data.user.UserDataSource
 import com.moonlightbutterfly.cryptohub.data.user.UserRepository
+import com.moonlightbutterfly.cryptohub.di.RepositoryModule
+import com.moonlightbutterfly.cryptohub.localpreferences.FakeLocalPreferencesDataSourceImpl
+import com.moonlightbutterfly.cryptohub.notifications.NotifierImpl
+import com.moonlightbutterfly.cryptohub.signin.FakeFirebaseSignInHandler
+import com.moonlightbutterfly.cryptohub.signin.FirebaseSignInHandler
+import com.moonlightbutterfly.cryptohub.user.FakeUserDataSourceImpl
 import dagger.Binds
 import dagger.Module
 import dagger.Provides
