+++ data/src/androidTest/java/com/moonlightbutterfly/cryptohub/user/FakeUserDataSourceImpl.kt
@@ -1,7 +1,8 @@
-package com.moonlightbutterfly.cryptohub.data.user
+package com.moonlightbutterfly.cryptohub.user
 
 import com.moonlightbutterfly.cryptohub.data.common.Answer
 import com.moonlightbutterfly.cryptohub.data.common.Error
+import com.moonlightbutterfly.cryptohub.data.user.UserDataSource
 import com.moonlightbutterfly.cryptohub.models.User
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flowOf
